requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
peft>=0.18.0  # SPECTER2模型需要
accelerate>=0.21.0  # peft依赖

# IR方法
rank-bm25>=0.2.2
scikit-learn>=1.3.0

# 向量搜索
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # 如果有GPU，取消注释这一行

# 机器学习
lightgbm>=3.3.0

# 工具
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# 实验跟踪 (可选)
# wandb>=0.15.0

# 超参数优化 (可选)
# optuna>=3.2.0

# 特征重要性分析 (可选)
# shap>=0.42.0

